requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
fake-useragent==0.1.11
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
lxml==4.5.0
numpy==1.18.4
pandas==1.0.3
parse==1.15.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-html==0.10.0
rsa==4.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
tqdm==4.46.0
urllib3==1.25.9
w3lib==1.21.0
websockets==8.1
